refactor: change loader on pages
