+++ src/router/router.tsx
@@ -5,10 +5,10 @@ import {
 } from 'react-router-dom';
 import RootLayout from '../layouts/RootLayout';
 import WelcomePage from '../pages/WelcomePage/WelcomePage';
-import SignInPage from '../pages/SignInPage';
-import SignUpPage from '../pages/SignUpPage';
+import SignInPage from '../pages/SignInPage/SignInPage';
+import SignUpPage from '../pages/SignUpPage/SignUpPage';
 import MainPage from '../pages/MainPage/MainPage';
-import NotFoundPage from '../pages/NotFoundPage';
+import NotFoundPage from '../pages/NotFoundPage/NotFoundPage';
 import AuthPrivateRoute from '../hoc/AuthPrivateRoute';
 
 const router = createBrowserRouter(
